+++ src/lm_polygraph/model_adapters/__init__.py
@@ -1,3 +1,4 @@
 from .whitebox_model_basic import WhiteboxModelBasic
+from .whitebox_model_vllm import WhiteboxModelvLLM
 from lm_polygraph.utils.model import WhiteboxModel
 from lm_polygraph.utils.model import BlackboxModel
